+++ src/component/detail/review/reviewBox.css
@@ -5,13 +5,14 @@
         box-shadow: 0px 0px 4px 0px #0000004d;
         border-radius: 15px;
         margin-bottom: 15px;
-        padding: 16px 18px;
+        /* padding: 16px 18px; */
     }
 
     .review_box_header {
         display: flex;
         flex-direction: row;
         align-items: center;
+        margin-bottom: 10px;
     }
 
     .review_box_header p {
@@ -27,23 +28,23 @@
     }
 
     .review_box_contents {
-        display: grid;
-        /* 한 줄에 세개의 열 */
-        grid-template-columns: repeat(3, 1fr);
-        gap: 5px;
+        /* display: grid;
+        grid-template-columns: repeat(3, 1fr); */
+        display: flex;
+        flex-wrap: wrap;
+        gap: 7px;
         margin-top: 7px;
-        /* width: 100%; */
     }
 
     .keyword_container {
+        display: inline-block;
         background-color: #9c9c9c26;
         border-radius: 5px;
-        border: none;
-        padding: 7px 10px;
-        font-size: 13px;
+        padding: 7px 9px;
+        font-size: 12px;
         font-weight: 400;
-        /* width: 30%;
-	flex: 1; */
+        margin-bottom: 2px;
+        max-width: calc(33.33%);
     }
 }
