+++ src/component/mypage/myReview.css
@@ -1,0 +1,148 @@
+@media screen and (min-width: 768px) {
+    .my_review_container {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        margin-top: 50px;
+        margin-bottom: 200px;
+    }
+
+    .my_review_container h2 {
+        font-size: 24px;
+        font-weight: 500px;
+        margin-bottom: 30px;
+    }
+
+    .my_review_box {
+        width: 475px;
+        background-color: white;
+        box-shadow: 0px 0px 4px 0px #0000004d;
+        border-radius: 15px;
+        margin-bottom: 15px;
+        padding: 16px 0px 12px 18px;
+    }
+
+    .review_comment {
+        font-size: 13px;
+        font-weight: 400;
+        padding: 5px 5px 0px 4px;
+    }
+
+    .club_name {
+        font-size: 14px;
+        font-weight: 700;
+        margin-right: 7px;
+        margin-bottom: 2px;
+        color: #7BC8E0;
+    }
+
+    .review_approved {
+        border: 1px solid #7BC8E0;
+        border-radius: 20px;
+        font-size: 10px;
+        font-weight: 600px;
+        padding: 3px 7px;
+        margin-left: 10px;
+        margin-bottom: 5px;
+        background-color: rgba(123, 200, 224, 0.15);
+        color: #7BC8E0;
+    }
+
+    .review_pending {
+        border: 1px solid #00000060;
+        border-radius: 20px;
+        font-size: 10px;
+        font-weight: 600px;
+        padding: 3px 7px;
+        margin-left: 10px;
+        margin-bottom: 5px;
+        background-color: #F2F2F2;
+        color: #00000060;
+    }
+
+    .review_rejected {
+        border: 1px solid #fd3c56b3;
+        border-radius: 20px;
+        font-size: 10px;
+        font-weight: 600px;
+        padding: 3px 7px;
+        margin-left: 10px;
+        margin-bottom: 5px;
+        background-color: #fd3c561f;
+        color: #fd3c56b3;
+    }
+}
+
+@media screen and (max-width: 768px) {
+    .my_review_container {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        margin-top: 50px;
+        margin-bottom: 200px;
+    }
+
+    .my_review_container h2 {
+        font-size: 24px;
+        font-weight: 500px;
+        margin-bottom: 30px;
+    }
+
+    .my_review_box {
+        width: 300px;
+        background-color: white;
+        box-shadow: 0px 0px 4px 0px #0000004d;
+        border-radius: 15px;
+        margin-bottom: 15px;
+        padding: 16px 18px;
+    }
+
+    .review_comment {
+        font-size: 13px;
+        font-weight: 400;
+        padding: 5px 5px 0px 4px;
+    }
+
+    .club_name {
+        font-size: 14px;
+        font-weight: 700;
+        margin-right: 7px;
+        color: #7BC8E0;
+    }
+
+    .review_approved {
+        border: 1px solid #7BC8E0;
+        border-radius: 20px;
+        font-size: 10px;
+        font-weight: 600px;
+        padding: 3px 7px;
+        margin-left: 10px;
+        margin-bottom: 5px;
+        background-color: rgba(123, 200, 224, 0.15);
+        color: #7BC8E0;
+    }
+
+    .review_pending {
+        border: 1px solid #00000060;
+        border-radius: 20px;
+        font-size: 10px;
+        font-weight: 600px;
+        padding: 3px 7px;
+        margin-left: 10px;
+        margin-bottom: 5px;
+        background-color: #F2F2F2;
+        color: #00000060;
+    }
+
+    .review_rejected {
+        border: 1px solid #fd3c56b3;
+        border-radius: 20px;
+        font-size: 10px;
+        font-weight: 600px;
+        padding: 3px 7px;
+        margin-left: 10px;
+        margin-bottom: 5px;
+        background-color: #fd3c561f;
+        color: #fd3c56b3;
+    }
+}
